DGen: Adding encounter tiles
One tile in a room (not island) is selected to
be mutated into an encounter. A Gaussian distribution
is used to select the tiles, which keeps them close to
the center of the rooms. These tiles are elevated and
are a medium grey in the minimap. It is also currently
possible for encounter tiles to be eroded away.