DGen/Combat: Adding switching to a combat scene for an encounter.
The combat prototype was added to main.blend and things were
hacked up to add the scene when an encounter was triggered.
Instead of restarting the game, the combat now just ends the scene.

The encounter distance was reduced to the point where the player
has to enter the same square as an encounter to trigger it.